Add ItemDetailsPage (buy items)

+++ M.A.G.U.S/Races/Amund.cs
@@ -5,13 +5,13 @@ namespace M.A.G.U.S.Races;
 
 public class Amund : Race
 {
-    public override short Strength => 1;
+    public override sbyte Strength => 1;
 
-    public override short Stamina => 1;
+    public override sbyte Stamina => 1;
 
-    public override short Beauty => 3;
+    public override sbyte Beauty => 3;
 
-    public override short Astral => -1;
+    public override sbyte Astral => -1;
 
     public override SpecialQualificationList SpecialQualifications =>
     [
