+++ M.A.G.U.S/Races/CourtGoblin.cs
@@ -9,17 +9,17 @@ namespace M.A.G.U.S.Races;
 /// </summary>
 public class CourtGoblin : Race
 {
-    public override short Strength => -2;
+    public override sbyte Strength => -2;
 
-    public override short Stamina => 1;
+    public override sbyte Stamina => 1;
 
-    public override short Intelligence => -1;
+    public override sbyte Intelligence => -1;
 
-    public override short Health => 1;
+    public override sbyte Health => 1;
 
-    public override short Beauty => -2;
+    public override sbyte Beauty => -2;
 
-    public override short Astral => -2;
+    public override sbyte Astral => -2;
 
     public override SpecialQualificationList SpecialQualifications =>
     [
