+++ M.A.G.U.S/Races/CourtOrc.cs
@@ -5,17 +5,17 @@ namespace M.A.G.U.S.Races;
 
 public class CourtOrc : Race
 {
-    public override short Strength => 2;
+    public override sbyte Strength => 2;
 
-    public override short Stamina => 1;
+    public override sbyte Stamina => 1;
 
-    public override short Health => 2;
+    public override sbyte Health => 2;
 
-    public override short Beauty => -3;
+    public override sbyte Beauty => -3;
 
-    public override short Intelligence => -1;
+    public override sbyte Intelligence => -1;
 
-    public override short Astral => -3;
+    public override sbyte Astral => -3;
 
     public override SpecialQualificationList SpecialQualifications =>
     [
