+++ M.A.G.U.S/Races/Goblin.cs
@@ -8,17 +8,17 @@ namespace M.A.G.U.S.Races;
 /// </summary>
 public class Goblin : Race
 {
-    public override short Strength => -2;
+    public override sbyte Strength => -2;
 
-    public override short Dexterity => 3;
+    public override sbyte Dexterity => 3;
 
-    public override short Speed => 3;
+    public override sbyte Speed => 3;
 
-    public override short Beauty => -4;
+    public override sbyte Beauty => -4;
 
-    public override short Willpower => -1;
+    public override sbyte Willpower => -1;
 
-    public override short Astral => -2;
+    public override sbyte Astral => -2;
 
     public override SpecialQualificationList SpecialQualifications =>
     [
