+++ M.A.G.U.S/Races/Khal.cs
@@ -8,19 +8,19 @@ namespace M.A.G.U.S.Races;
 
 public class Khal : Race
 {
-    public override short Strength => 3;
+    public override sbyte Strength => 3;
 
-    public override short Speed => 2;
+    public override sbyte Speed => 2;
 
-    public override short Dexterity => 1;
+    public override sbyte Dexterity => 1;
 
-    public override short Stamina => 2;
+    public override sbyte Stamina => 2;
 
-    public override short Health => 3;
+    public override sbyte Health => 3;
 
-    public override short Willpower => -1;
+    public override sbyte Willpower => -1;
 
-    public override short Astral => -5;
+    public override sbyte Astral => -5;
 
     public override QualificationList Qualifications =>
     [
