+++ M.A.G.U.S/Races/Wier.cs
@@ -5,9 +5,9 @@ namespace M.A.G.U.S.Races;
 
 public class Wier : Race
 {
-    public override short Beauty => 1;
+    public override sbyte Beauty => 1;
 
-    public override short Intelligence => 1;
+    public override sbyte Intelligence => 1;
 
     public override SpecialQualificationList SpecialQualifications =>
     [
